Create custom hook useIsAdult
